modified CreateAnnouncementTest.kt accordingly to the new notification screen and moved it into announcement package. Also deleted AnnouncementViewModelTest.kt which was obsolete
